fix sorting state propagating to table
